pyproject.toml
[build-system]
build-backend = 'flit_core.buildapi'
requires = ['flit_core >=3.4,<4']

[project]
authors = [
  {name = 'Matthieu Leclair', email = 'matthieu.leclair@env.ethz.ch'},
  {name = 'Julian Geiger', email = 'julian.geiger@psi.ch'},
  {name = 'Alexander Goscinski', email = 'alexander.goscinski@psi.ch'},
]

requires-python = '>=3.9'
dependencies = [
    'aiida-core~=2.5',
    'aiida-workgraph',
    'isoduration',
    'PyYAML'
]

classifiers = [
  'Development Status :: 1 - Planning',
  'License :: OSI Approved :: MIT License',
  'Operating System :: POSIX :: Linux',
  'Programming Language :: Python',
  'Programming Language :: Python :: 3.9',
  'Programming Language :: Python :: 3.10',
  'Programming Language :: Python :: 3.11',
  'Programming Language :: Python :: 3.12',
  'Topic :: Scientific/Engineering',
  'Topic :: Scientific/Engineering :: Atmospheric Science',
]
description = 'WcFlow a library for creating weather and climate workflows.'
keywords = ['aiida', 'workflows']
name = 'wcflow'
readme = 'README.org'
dynamic = ["version"] # read from __init__py

[project.optional-dependencies]
tests = [
    'pytest~=7.0',
]
